New defaults and added a speed up feature
Changed the default bursts to display to 100 and increased the maximum burst radius to 150. I think this is a better default for fly testing. I also added a condition that increased the burst speed by 1 after every 5 bursts (this can cause a strobe-like effect past speed 15 so I might consider altering this)