Add additional rows (broken)
Adds a button that adds more rows to the metadata input form. However, after this button is pressed for the first time, the eventListeners for the start, stop, and restart button stop working. The listener to add more rows continues to work, though, so the issue only impacts buttons in the 'rightBox' div.